+++ RELEASE_NOTES.md
@@ -1,10 +1,11 @@
 ## Minor Updates
 
-Notes update.
+Introduction of Notes to the application.
 
 - Added the ability to write a note per day
+- Notes are encrypted per user, and also requires a new environment variable `NOTE_ENCRYPTION_KEY` to be set
 - Added stats for notes alongside actions
 - New API endpoints for the notes feature
 - Some UI updates to make things fit and flow smoothly
 - Added user setting to define note colour for calendar highlight
-- Storing current note draft between application navigation
+- Storing current day's note as a draft between application navigation if the user hasn't saved it